Create BaseNeuron class, which takes care of common operations between miner and validator neurons. Also move everything in sync.py to the respective neuron classes and delete
