+++ .circleci/config.yml
@@ -42,6 +42,42 @@ jobs:
             . env/bin/activate
             python -m black --exclude '(env|venv|.eggs)' --check .
 
+    pylint:
+      resource_class: small
+      parameters:
+        python-version:
+          type: string
+      docker:
+        - image: cimg/python:<< parameters.python-version >>
+
+      steps:
+        - checkout
+
+        - restore_cache:
+            name: Restore cached pylint venv
+            keys:
+              - v1-pypi-py-pylint-<< parameters.python-version >>
+
+        - run:
+            name: Update & Activate pylint venv
+            command: |
+              python -m venv env/
+              . env/bin/activate
+              python -m pip install --upgrade pip
+              pip install pylint
+
+        - save_cache:
+            name: Save cached pylint venv
+            paths:
+              - "env/"
+            key: v1-pypi-py-pylint-<< parameters.python-version >>
+
+        - run:
+            name: Pylint check
+            command: |
+              . env/bin/activate
+              python -m pylint --rcfile .pylintrc bittensor/
+
   check_compatibility:
     parameters:
       python_version:
@@ -131,10 +167,12 @@ workflows:
       - check_compatibility:
           python_version: "3.11"
           name: check-compatibility-3.11
-  
+
   pr-requirements:
     jobs:
       - black:
+          python-version: "3.8.12"
+      - pylint:
           python-version: "3.8.12"
       - build:
           matrix:
