+++ template/base/neuron.py
@@ -1,0 +1,168 @@
+# The MIT License (MIT)
+# Copyright © 2023 Yuma Rao
+# TODO(developer): Set your name
+# Copyright © 2023 <your name>
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+# documentation files (the “Software”), to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all copies or substantial portions of
+# the Software.
+
+# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
+# THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+# DEALINGS IN THE SOFTWARE.
+
+import copy
+import typing
+
+import bittensor as bt
+
+from abc import ABC, abstractmethod
+
+# Sync calls set weights and also resyncs the metagraph.
+from template.utils.sync import check_registered
+from template.utils.config import check_config, add_args, config
+from template.utils.misc import ttl_get_block
+
+
+class BaseNeuron(ABC):
+    """
+    Base class for Bittensor miners.
+    """
+
+    @classmethod
+    def check_config(cls, config: "bt.Config"):
+        check_config(cls, config)
+
+    @classmethod
+    def add_args(cls, parser):
+        add_args(cls, parser)
+
+    @classmethod
+    def config(cls):
+        return config(cls)
+
+    subtensor: "bt.subtensor"
+    wallet: "bt.wallet"
+    metagraph: "bt.metagraph"
+
+    @property
+    def block(self):
+        return ttl_get_block(self)
+
+    def __init__(self, config=None):
+        base_config = copy.deepcopy(config or BaseNeuron.config())
+        self.config = self.config()
+        self.config.merge(base_config)
+        self.check_config(self.config)
+
+        # Set up logging with the provided configuration and directory.
+        bt.logging(config=self.config, logging_dir=self.config.full_path)
+        bt.logging.info(
+            f"Running validator for subnet: {self.config.netuid} on network: {self.subtensor.chain_endpoint}"
+        )
+
+        # If a gpu is required, set the device to cuda:N (e.g. cuda:0)
+        self.device = self.config.neuron.device
+
+        # Log the configuration for reference.
+        bt.logging.info(self.config)
+
+        # Build Bittensor objects
+        # These are core Bittensor classes to interact with the network.
+        bt.logging.info("Setting up bittensor objects.")
+
+        # The wallet holds the cryptographic key pairs for the miner.
+        self.wallet = bt.wallet(config=self.config)
+        bt.logging.info(f"Wallet: {self.wallet}")
+
+        # The subtensor is our connection to the Bittensor blockchain.
+        self.subtensor = bt.subtensor(config=self.config)
+        bt.logging.info(f"Subtensor: {self.subtensor}")
+
+        # The metagraph holds the state of the network, letting us know about other validators and miners.
+        self.metagraph = self.subtensor.metagraph(self.config.netuid)
+        bt.logging.info(f"Metagraph: {self.metagraph}")
+
+        # Check if the miner is registered on the Bittensor network before proceeding further.
+        self.check_registered()
+
+        # Each miner gets a unique identity (UID) in the network for differentiation.
+        self.uid = self.metagraph.hotkeys.index(self.wallet.hotkey.ss58_address)
+        bt.logging.info(f"Running miner on uid: {self.uid}")
+
+        self.step = 0
+
+    @abstractmethod
+    async def forward(self, synapse: bt.Synapse) -> bt.Synapse:
+        ...
+
+    # async def backward(self, synapse: bt.Synapse) -> bt.Synapse:
+    #     pass
+
+    @abstractmethod
+    async def blacklist(self, synapse: bt.Synapse) -> typing.Tuple[bool, str]:
+        ...
+
+    @abstractmethod
+    async def priority(self, synapse: bt.Synapse) -> float:
+        ...
+
+    @abstractmethod
+    def run(self):
+        ...
+
+    def sync(self):
+        """
+        Wrapper for synchronizing the state of the network for the given miner or validator.
+        """
+        # Ensure miner or validator hotkey is still registered on the network.
+        self.check_registered()
+
+        if self.should_sync_metagraph():
+            self.resync_metagraph()
+
+        if self.should_set_weights():
+            self.set_weights()
+
+        # Always save state.
+        self.save_state()
+
+    def check_registered(self):
+        # --- Check for registration.
+        if not self.subtensor.is_hotkey_registered(
+            netuid=self.config.netuid,
+            hotkey_ss58=self.wallet.hotkey.ss58_address,
+        ):
+            bt.logging.error(
+                f"Wallet: {self.wallet} is not registered on netuid {self.config.netuid}"
+                f"Please register the hotkey using `btcli subnets register` before trying again"
+            )
+            exit()
+
+    def should_sync_metagraph(self):
+        """
+        Check if enough epoch blocks have elapsed since the last checkpoint to sync.
+        """
+        return (
+            self.block - self.metagraph.last_update[self.uid]
+        ) > self.config.neuron.epoch_length
+
+    def should_set_weights(self) -> bool:
+        # Check if enough epoch blocks have elapsed since the last epoch.
+        if self.config.neuron.disable_set_weights:
+            return False
+
+        return (
+            self.block - self.metagraph.last_update[self.uid]
+        ) > self.config.neuron.epoch_length
+
+    def save_state(self):
+        bt.logging.warning(
+            "save_state() not implemented for this neuron. You can implement this function to save model checkpoints or other useful data."
+        )
